+++ Sources/Sentry/SentryReplaySettings.m
@@ -21,8 +21,8 @@ SentryReplaySettings ()
     return self;
 }
 
-- (instancetype)initWithReplaySessionSampleRate:(CGFloat)sessionSampleRate
-                       replaysOnErrorSampleRate:(CGFloat)errorSampleRate
+- (instancetype)initWithReplaySessionSampleRate:(float)sessionSampleRate
+                       replaysOnErrorSampleRate:(float)errorSampleRate
 {
     if (self = [self init]) {
         self.replaysSessionSampleRate = sessionSampleRate;
